pyproject.toml
[project]
name = "fastapitutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.129.0",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "imagekitio>=5.2.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]


